rename(route-folder): chatroom types 폴더 분리
