+++ src/app/(route)/chat/[postId]/_components/ChatRoomMain/_internal/ExpandableMessageBubble/ExpandableMessageBubble.tsx
