+++ src/app/(route)/chat/[postId]/_components/InputChat/_internal/InputChatImageSection.tsx
@@ -2,11 +2,11 @@
 
 import { Icon, ImageSelectButton } from "@/components";
 import useSendImage from "@/api/fetch/chatMessage/api/useSendImage";
-import { useHandleSendImage } from "./useHandleSendImage";
 import {
   InputChatImageSectionIds,
   InputChatImageSectionImageState,
-} from "./InputChatImageSectionType";
+} from "../../../_types/InputChatImageSection";
+import { useHandleSendImage } from "../../../_hooks/useHandleSendImage/useHandleSendImage";
 
 interface InputChatImageSectionProps {
   ids: InputChatImageSectionIds;
