+++ src/app/(route)/chat/[postId]/_hooks/useHandleSendImage/useHandleSendImage.ts
