+++ src/app/(route)/chat/[postId]/_types/InfoButtonOptionValue.ts
