+++ src/app/(route)/chat/[postId]/_types/InputChatImageSection.ts
